Moved things out of the service into the core. Changed function outputs to be schema entries instead of the ares data schema struct.
