+++ Ares.Core/Ares.Core.csproj
@@ -18,7 +18,7 @@
 
 
   <ItemGroup>
-	  <PackageReference Include="ARES.Datamodel" Version="0.17.0-alpha.8" />
+	  <PackageReference Include="ARES.Datamodel" Version="0.17.0-alpha.13" />
 	  <PackageReference Include="CsvHelper" Version="33.1.0" />
 	  <PackageReference Include="DynamicData" Version="9.4.1" />
 	  <PackageReference Include="Grpc.Net.Client" Version="2.76.0" />
