+++ Ares.Device/AresDevicePluginBase/Ares.Device.csproj
@@ -35,7 +35,7 @@
   </PropertyGroup>
 
   <ItemGroup>
-    <PackageReference Include="ARES.Datamodel" Version="0.17.0-alpha.8" />
+    <PackageReference Include="ARES.Datamodel" Version="0.17.0-alpha.13" />
     <PackageReference Include="System.Reactive" Version="6.1.0" />
   </ItemGroup>
