+++ AresMessaging/AresMessaging.csproj
@@ -26,7 +26,7 @@
   </ItemGroup>
 
   <ItemGroup>
-    <PackageReference Include="ARES.Datamodel" Version="0.17.0-alpha.8" GeneratePathProperty="True" />
+    <PackageReference Include="ARES.Datamodel" Version="0.17.0-alpha.13" GeneratePathProperty="True" />
     <PackageReference Include="Google.Protobuf" Version="3.33.3" />
     <PackageReference Include="Grpc" Version="2.46.6" />
     <PackageReference Include="Grpc.Tools" Version="2.76.0">
