+++ AresScript/AresSystemFunction.cs
@@ -4,6 +4,6 @@ namespace AresScript;
 
 public delegate Task<AresValue> AresFunctionDelegate(List<AresValue> args, CancellationToken token);
 
-public record AresSystemFunction(string Id, string Name, AresFunctionDelegate Body, AresDataSchema InputSchema, AresDataSchema OutputSchema, string Namespace = "", string Description = "")
+public record AresSystemFunction(string Id, string Name, AresFunctionDelegate Body, AresDataSchema InputSchema, SchemaEntry OutputSchema, string Namespace = "", string Description = "")
 {
 }
