+++ DemoRemoteAnalyzer/DemoRemoteAnalyzer.csproj
@@ -7,7 +7,7 @@
   </PropertyGroup>
 
   <ItemGroup>
-    <PackageReference Include="ARES.Datamodel" Version="0.17.0-alpha.8" />
+    <PackageReference Include="ARES.Datamodel" Version="0.17.0-alpha.13" />
     <PackageReference Include="Grpc.AspNetCore" Version="2.76.0" />
   </ItemGroup>
