+++ RestDevice/RestDevice.csproj
@@ -17,7 +17,7 @@
 	</PropertyGroup>
 
   <ItemGroup>
-    <PackageReference Include="ARES.Datamodel" Version="0.17.0-alpha.8" GeneratePathProperty="True" />
+    <PackageReference Include="ARES.Datamodel" Version="0.17.0-alpha.13" GeneratePathProperty="True" />
     <PackageReference Include="Grpc.Tools" Version="2.76.0">
       <PrivateAssets>all</PrivateAssets>
       <IncludeAssets>runtime; build; native; contentfiles; analyzers; buildtransitive</IncludeAssets>
